Move SqlInvokedRegularFunctionTestUtils to presto-sql-functions
Also rename the class to SqlInvokedFunctionTestUtils
